reintegrate fix for sf bug #3111545: double instead of float in aspect ration documentation
git-svn-id: https://svn.code.sf.net/p/auber/code/branches/tulip_3_4_maint@3302 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ plugins/layout/SquarifiedTreeMap.cpp
@@ -17,7 +17,6 @@
  *
  */
 #include <cmath>
-#include <float.h>
 #include <algorithm>
 #include <tulip/TreeTest.h>
 #include "SquarifiedTreeMap.h"
@@ -52,7 +51,7 @@ namespace
       HTML_HELP_CLOSE(),
       // aspect ratio
       HTML_HELP_OPEN()				 \
-      HTML_HELP_DEF( "type", "float" ) \
+      HTML_HELP_DEF( "type", "double" ) \
       HTML_HELP_DEF( "default", "1." )	 \
       HTML_HELP_BODY() \
       "This parameter enables to set up the aspect ratio (height/width) for the rectangle corresponding to the root node." \
